Move store rehydrate to `index.tsx`
State must be consistent to avoid incorrect updates to state. Doing this
at a component level introduces many separate storage listeners. Pulling
this listener up to `index.tsx` ensures all stores are updated as
required and the number of listeners per window is 1.
