+++ src/app/pages/recording-source-selector.tsx
@@ -7,10 +7,8 @@ import {
   expandRecordingSourceSelector,
 } from "../../api/windows";
 import RecordingSource from "../../features/recording-source/components/recording-source";
-import { rehydrateRecordingPreferencesStore } from "../../stores/recording-preferences.store";
 import {
   AppWindow,
-  rehydrateWindowReopenState,
   useWindowReopenStore,
 } from "../../stores/window-open-state.store";
 import { Events } from "../../types/events";
@@ -21,11 +19,6 @@ const RecordingSourceSelector = () => {
   );
 
   const [isExpanded, setIsExpanded] = useState(false);
-
-  const rehydrateStores = (e: StorageEvent) => {
-    rehydrateWindowReopenState(e);
-    rehydrateRecordingPreferencesStore(e);
-  };
 
   const onToggle = () => {
     // Invert the state
@@ -49,13 +42,6 @@ const RecordingSourceSelector = () => {
   useEffect(() => {
     setIsExpanded(false);
   }, [startRecordingDockOpened]);
-
-  useEffect(() => {
-    window.addEventListener("storage", rehydrateStores);
-    return () => {
-      window.removeEventListener("storage", rehydrateStores);
-    };
-  }, []);
 
   return (
     <div className="flex flex-row p-2 h-[100vh] w-full items-end justify-center">
